self-healing/api/requirements.txt
# Self-Healing System API Dependencies
# API服务依赖包列表

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# 异步支持
aiofiles==23.2.1
aiohttp==3.9.1
aioredis==2.0.1
asyncio-mqtt==0.16.1

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0  # PostgreSQL异步驱动

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 监控和指标
prometheus-client==0.19.0
psutil==5.9.6

# 日志
loguru==0.7.2
structlog==23.2.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 任务队列
celery==5.3.4
redis==5.0.1

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# JSON处理
orjson==3.9.10

# 模板引擎
jinja2==3.1.2

# 文件处理
watchdog==3.0.0

# 网络工具
ping3==4.0.4
netifaces==0.11.0

# 系统工具
click==8.1.7
rich==13.7.0

# 测试工具（开发依赖）
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # 用于测试HTTP客户端
factory-boy==3.3.0
faker==20.1.0

# 代码质量
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8

# 部署工具
gunicorn==21.2.0

# 环境变量
python-decouple==3.8